+++ CHANGES.md
@@ -4,6 +4,7 @@ Changes
 1.0.11-SNAPSHOT
 -----------------------
 
+* Added support of Flash Socket transport
 * Added `SocketIOAdapter` class
 
 1.0.10 / 2013-12-24